Refactor/snapshot

+++ src/RewardFlow.API/Common/EmployeeLookup/EmployeeSalaryDto.cs
@@ -3,5 +3,5 @@ namespace Reward_Flow_v2.Common.EmployeeLookup;
 public record EmployeeSalaryDto
 {
     public int EmployeeId { get; init; }
-    public float Salary { get; init; }
+    public decimal Salary { get; init; }
 }
